# Optimize parsing and conversion in DurationConverter, reduce allocations