--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Duration.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Duration.cs
@@ -481,18 +481,22 @@ namespace System.Windows
         /// <returns>A string representation of this Duration.</returns>
         public override string ToString()
         {
-            if (HasTimeSpan)
-            {
-                return _timeSpan.ToString();
-            }
-            else if (_durationType == DurationType.Forever)
-            {
+            return HasTimeSpan ? TypeDescriptor.GetConverter(_timeSpan).ConvertToString(_timeSpan) : ToStringInvariant();
+        }
+
+        /// <summary>
+        /// This method does not use the current <see cref="System.TimeSpan"/>'s <see cref="TypeDescriptor"/> for retrieving the
+        /// current converter, but rather uses the standard <see cref="TimeSpan.ToString()"/> override for <see cref="DurationConverter"/> needs.
+        /// </summary>
+        /// <returns>A culture-invariant representation of the <see cref="Duration"/> instance.</returns>
+        internal string ToStringInvariant()
+        {
+            if (_durationType == DurationType.Forever)
                 return "Forever";
-            }
-            else // IsAutomatic
-            {
+            else if (_durationType == DurationType.Automatic)
                 return "Automatic";
-            }
+
+            return _timeSpan.ToString();
         }
 
         #endregion