--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/DurationConverter.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/DurationConverter.cs
@@ -22,14 +22,7 @@ namespace System.Windows
         /// <ExternalAPI/>
         public override bool CanConvertFrom(ITypeDescriptorContext td, Type t)
         {
-            if (t == typeof(string))
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
+            return t == typeof(string);
         }
 
         /// <summary>